requirements.txt
yoyo-migrations==5.0.5
requests
pylint
pyyaml
